Add linting to the eslint-config-hmpps package and rename rules folder to make it clear it only houses Airbnb rule sets More dogfooding, except in Airbnb folder as those files are taken as-is
